func_app/requirements.txt
pyzabbix==1.3.0
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
requests==2.31.0
azure-functions==1.18.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
python-dotenv>=1.0.0